requirements.txt
torch>=1.10
torchvision>=0.11
opencv-python
Pillow
numpy
